backend/src/requirements.txt
asgiref==3.8.1
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
pillow==10.3.0
PyJWT==2.8.0
PyMySQL==1.1.0
sqlparse==0.5.0
typing_extensions==4.11.0
